databricks_pipeline/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
argparse
pathlib
apify-client>=1.12.2
openpyxl
instaloader
google-genai